Скорректировал действия по модулям в сетке программы
