+++ keyboards/inline/calendar_inline/__init__.py
@@ -1,0 +1,1 @@
+from . import inline_calendar
