+++ keyboards/inline/calendar_inline/inline_calendar.py
@@ -1,6 +1,7 @@
 import calendar
 from datetime import date
 from telebot.types import InlineKeyboardButton, InlineKeyboardMarkup
+from loader import bot
 
 
 def create_callback_data(action, year, month, day):
@@ -13,6 +14,38 @@ def create_callback_data(action, year, month, day):
     :return: Строку для callback_data
     """
     return ";".join([action, str(year), str(month), str(day)])
+
+
+def get_list_data(data):
+    """
+    Делает список для callback
+    :param data:
+    :return:
+    """
+    return data.split(';')
+
+
+def get_next_or_prev_mont(action, year, month):
+    """
+    Функция корректирует год и месяц если пользователь нажал на стрелочки выбора месяца
+    :param action: След или предыдущий месяц
+    :param year: год с коллбэка
+    :param month: месяц оттудаже
+    :return: клавиатуру с новыми данными
+    """
+    if action == 'next':
+        if month == 12:
+            year += 1
+            month = 1
+        else:
+            month += 1
+    else:
+        if month == 1:
+            year -= 1
+            month = 12
+        else:
+            month -= 1
+    return bot_get_keyboard_inline(year=year, month=month)
 
 
 def bot_get_keyboard_inline(year=None, month=None) -> InlineKeyboardMarkup:
@@ -47,3 +80,20 @@ def bot_get_keyboard_inline(year=None, month=None) -> InlineKeyboardMarkup:
         InlineKeyboardButton('>>', callback_data=create_callback_data('Next-month', year, month, day=1)),
     )
     return keyboard
+
+
+@bot.callback_query_handler(func=lambda call: not call.data.startswith('DAY'))
+def callback_inline(call):
+    """
+    Ловит выбор пользователя с календаря, если было выбрано перелистывание месяца
+    :param call: Выбор пользователя
+    """
+    if call.data.startswith('EMPTY'):
+        bot.answer_callback_query(callback_query_id=call.id,
+                                  text='Выберите число!')  # Применить, когда тыкает в ненужное место
+    else:
+        date = get_list_data(call.data)
+        action = 'prev' if call.data.startswith('Prev') else 'next'
+        bot.edit_message_text('Месяц', call.message.chat.id, call.message.id,
+                              reply_markup=get_next_or_prev_mont(action=action, year=int(date[1]),
+                                                                 month=int(date[2])))
